Crate cairo_native_runtime

Structs

DICT_GAS_REFUND_PER_ACCESS

DictValuesArrayAbi

Needed for the correct alignment, since the key [u8; 32] in rust has 8 byte alignment but its a felt, so in reality it has 16.

FormattedItem

A formatted string representation of anything formattable (e.g. ByteArray, felt, short-string).

HALF_PRIME

Constants

BYTES_IN_WORD

BYTE_ARRAY_MAGIC

Functions

as_cairo_short_string

Converts a bigint representing a felt252 to a Cairo short-string.

as_cairo_short_string_ex

Converts a bigint representing a felt252 to a Cairo short-string of the given length. Nulls are allowed and length must be <= 31.

cairo_native__alloc_dict⚠

Allocates a new dictionary. Internally a rust hashmap: HashMap<[u8; 32], NonNull<()>

cairo_native__dict_free⚠

Frees the dictionary.

cairo_native__dict_gas_refund⚠

Compute the total gas refund for the dictionary at squash time.

cairo_native__dict_get⚠

Gets the value for a given key, the returned pointer is null if not found. Increments the access count.

cairo_native__dict_insert⚠

Inserts the provided key value. Returning the old one or nullptr if there was none.

cairo_native__dict_values⚠

Returns a array over the values of the dict, used for deep cloning.

cairo_native__libfunc__debug__print⚠

Based on cairo-lang-runner’s implementation.

cairo_native__libfunc__ec__ec_point_from_x_nz⚠

Compute ec_point_from_x_nz(x) and store it.

cairo_native__libfunc__ec__ec_point_try_new_nz⚠

Compute ec_point_try_new_nz(x).

cairo_native__libfunc__ec__ec_state_add⚠

Compute ec_state_add(state, point) and store the state back.

cairo_native__libfunc__ec__ec_state_add_mul⚠

Compute ec_state_add_mul(state, scalar, point) and store the state back.

cairo_native__libfunc__ec__ec_state_init⚠

Compute ec_state_init() and store the state back.

cairo_native__libfunc__ec__ec_state_try_finalize_nz⚠

Compute ec_state_try_finalize_nz(state) and store the result.

cairo_native__libfunc__hades_permutation⚠

Compute hades_permutation(op0, op1, op2) and replace the operands with the results.

cairo_native__libfunc__pedersen⚠

Compute pedersen(lhs, rhs) and store it into dst.

format_next_item

Formats a string or a short string / felt252. Returns the formatted string and a boolean indicating whether it’s a string. If can’t format the item, returns None.

Type Aliases

FeltDict

Felt252 type used in cairo native runtime