Crate cairo_native_runtime

Structs

DICT_GAS_REFUND_PER_ACCESS

HALF_PRIME

Functions

cairo_native__alloc_dict⚠

Allocates a new dictionary. Internally a rust hashmap: HashMap<[u8; 32], NonNull<()>

cairo_native__dict_free⚠

Frees the dictionary.

cairo_native__dict_gas_refund⚠

Compute the total gas refund for the dictionary at squash time.

cairo_native__dict_get⚠

Gets the value for a given key, the returned pointer is null if not found. Increments the access count.

cairo_native__dict_insert⚠

Inserts the provided key value. Returning the old one or nullptr if there was none.

cairo_native__libfunc__debug__print⚠

Based on cairo-lang-runner’s implementation.

cairo_native__libfunc__ec__ec_point_from_x_nz⚠

Compute ec_point_from_x_nz(x) and store it.

cairo_native__libfunc__ec__ec_point_try_new_nz⚠

Compute ec_point_try_new_nz(x).

cairo_native__libfunc__ec__ec_state_add⚠

Compute ec_state_add(state, point) and store the state back.

cairo_native__libfunc__ec__ec_state_add_mul⚠

Compute ec_state_add_mul(state, scalar, point) and store the state back.

cairo_native__libfunc__ec__ec_state_try_finalize_nz⚠

Compute ec_state_try_finalize_nz(state) and store the result.

cairo_native__libfunc__hades_permutation⚠

Compute hades_permutation(op0, op1, op2) and replace the operands with the results.

cairo_native__libfunc__pedersen⚠

Compute pedersen(lhs, rhs) and store it into dst.